old-tests/submission/Infraware/Forms/contents/Forms/button_formnovalidate.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_formNoValidate</h3>
	</p>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><button id='button_id'>button</button></p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		button.formNoValidate = true;

		if (typeof(button.formNoValidate) == "boolean") {
			test(function() {
				assert_equals(button.formNoValidate, true, "formNoValidate attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formNoValidate attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
